requirements.txt
crepe
librosa
matplotlib
music21
numpy
Pillow
pyaudio
soundfile





